requirements.txt
selenium~=3.141.0
airtest-selenium
airtest
pytest~=6.1.2
pytest-html
pytest-rerunfailures
unittest2
PyMySQL~=0.10.1
pywin32
pyyaml~=5.3.1
configparser~=5.0.1
pytest-assume
pytest-xdist
pandas
numpy
pocoui
pocounit
allure-pytest
requests~=2.23.0
redis~=3.5.3
rootpath~=0.1.1